+++ packages/medusa-plugin-mailchimp/src/api/index.js
@@ -1,0 +1,10 @@
+import { Router } from "express"
+import routes from "./routes"
+
+export default (rootDirectory) => {
+  const app = Router()
+
+  routes(app, rootDirectory)
+
+  return app
+}