+++ packages/medusa/src/api/routes/store/shipping-options/index.js
@@ -6,6 +6,7 @@ const route = Router()
 export default app => {
   app.use("/shipping-options", route)
 
+  route.get("/", middlewares.wrap(require("./list-options").default))
   route.get(
     "/:cart_id",
     middlewares.wrap(require("./list-shipping-options").default)