+++ packages/medusa/src/services/discount.js
@@ -455,7 +455,8 @@ class DiscountService extends BaseService {
       )
     }
 
-    return decorated
+    const final = await this.runDecorators_(decorated)
+    return final
   }
 }