+++ packages/medusa/src/services/line-item.js
@@ -1,5 +1,6 @@
 import { Validator, MedusaError } from "medusa-core-utils"
 import { BaseService } from "medusa-interfaces"
+import _ from "lodash"
 
 /**
  * Provides layer to manipulate line items.
@@ -38,6 +39,7 @@ class LineItemService extends BaseService {
     const lineItemSchema = Validator.object({
       title: Validator.string().required(),
       is_giftcard: Validator.bool().optional(),
+      should_merge: Validator.bool().optional(),
       description: Validator.string()
         .allow("")
         .optional(),
@@ -110,6 +112,7 @@ class LineItemService extends BaseService {
       title: product.title,
       description: variant.title,
       quantity,
+      should_merge: true,
       thumbnail: product.thumbnail,
       content: {
         unit_price,
@@ -117,11 +120,13 @@ class LineItemService extends BaseService {
         product,
         quantity: 1,
       },
+      metadata: {
+        ...metadata,
+      },
     }
 
     if (product.is_giftcard) {
       line.is_giftcard = true
-      line.metadata = metadata
     }
 
     return line
@@ -142,7 +147,8 @@ class LineItemService extends BaseService {
     } else if (!Array.isArray(match.content)) {
       return (
         line.content.variant._id.equals(match.content.variant._id) &&
-        line.content.quantity === match.content.quantity
+        line.content.quantity === match.content.quantity &&
+        _.isEqual(line.metadata, match.metadata)
       )
     }