+++ packages/medusa-plugin-slack-notification/CHANGELOG.md
@@ -3,6 +3,96 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.11 (2020-10-19)
+
+
+
+## 1.0.10 (2020-09-09)
+
+
+### Bug Fixes
+
+* updates license ([db519fb](https://github.com/medusajs/medusa/commit/db519fbaa6f8ad02c19cbecba5d4f28ba1ee81aa))
+
+
+
+## 1.0.7 (2020-09-07)
+
+
+
+## 1.0.5 (2020-09-06)
+
+
+### Bug Fixes
+
+* **medusa-plugin-slack-notification:** adds currency code to line items ([83c49e0](https://github.com/medusajs/medusa/commit/83c49e0b9cf0c6bc30aefc47fcde10ed221b89e3))
+
+
+
+## 1.0.4 (2020-09-06)
+
+
+
+## 1.0.3 (2020-09-06)
+
+
+### Bug Fixes
+
+* **medusa-plugin-slack-order-notification:** adds additional information ([#102](https://github.com/medusajs/medusa/issues/102)) ([f732a2d](https://github.com/medusajs/medusa/commit/f732a2d26a3ead4ef330a555961b49222e1f7244))
+
+
+
+## 1.0.1 (2020-09-05)
+
+
+
+## 1.0.1-beta.0 (2020-09-04)
+
+
+### Bug Fixes
+
+* **medusa-plugin-slack-notification:** Fixes slackbot ([#100](https://github.com/medusajs/medusa/issues/100)) ([1c8384c](https://github.com/medusajs/medusa/commit/1c8384c7c342ac3cb703f07a6d4d705b327e9716))
+
+
+
+# 1.0.0 (2020-09-03)
+
+
+
+# 1.0.0-alpha.30 (2020-08-28)
+
+
+
+# 1.0.0-alpha.27 (2020-08-27)
+
+
+
+# 1.0.0-alpha.26 (2020-08-27)
+
+
+
+# 1.0.0-alpha.24 (2020-08-27)
+
+
+
+# 1.0.0-alpha.3 (2020-08-20)
+
+
+
+# 1.0.0-alpha.2 (2020-08-20)
+
+
+
+# 1.0.0-alpha.1 (2020-08-20)
+
+
+
+# 1.0.0-alpha.0 (2020-08-20)
+
+
+
+
+
 ## [1.0.10](https://github.com/medusajs/medusa/compare/v1.0.9...v1.0.10) (2020-09-09)